Fix networking pages not checking state if stationary
